user-related models moved to user directory
