+++ app/src/main/java/com/nailorsh/repeton/common/data/models/Subject.kt
@@ -1,6 +1,6 @@
 package com.nailorsh.repeton.common.data.models
 
 data class Subject(
-    val id: Int,
-    val subjectName: String
+    val id : String,
+    val name: String
 )
