+++ app/src/main/java/com/nailorsh/repeton/common/data/models/user/Country.kt
@@ -1,0 +1,6 @@
+package com.nailorsh.repeton.common.data.models.user
+
+data class Country(
+    val id : String,
+    val name : String,
+)
