+++ app/src/main/java/com/nailorsh/repeton/common/data/models/user/Location.kt
@@ -1,0 +1,7 @@
+package com.nailorsh.repeton.common.data.models.user
+
+data class Location(
+    val id : String,
+    val country : Country,
+    val city : City,
+)
