+++ app/src/main/java/com/nailorsh/repeton/features/navigation/routes/TutorViewScreen.kt
@@ -1,6 +1,6 @@
 package com.nailorsh.repeton.features.navigation.routes
 
-import com.nailorsh.repeton.common.data.models.UserId
+import com.nailorsh.repeton.common.data.models.user.UserId
 import com.nailorsh.repeton.core.navigation.NavigationRoute
 
 sealed interface TutorViewScreen : NavigationRoute {
