+++ app/src/main/java/com/nailorsh/repeton/features/tutorprofile/data/TutorProfileRepository.kt
@@ -1,7 +1,7 @@
 package com.nailorsh.repeton.features.tutorprofile.data
 
-import com.nailorsh.repeton.common.data.models.Tutor
-import com.nailorsh.repeton.common.data.models.UserId
+import com.nailorsh.repeton.common.data.models.user.Tutor
+import com.nailorsh.repeton.common.data.models.user.UserId
 
 interface TutorProfileRepository {
     suspend fun getTutorProfile(id: UserId): Tutor
