+++ app/src/main/java/com/nailorsh/repeton/features/tutorsearch/data/TutorSearchRepository.kt
@@ -1,6 +1,6 @@
 package com.nailorsh.repeton.features.tutorsearch.data
 
-import com.nailorsh.repeton.common.data.models.Tutor
+import com.nailorsh.repeton.common.data.models.user.Tutor
 
 interface TutorSearchRepository {
     suspend fun getTutors(): List<Tutor>
