+++ app/src/main/java/com/nailorsh/repeton/features/tutorsearch/presentation/viewmodel/TutorSearchViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.nailorsh.repeton.common.data.models.Tutor
+import com.nailorsh.repeton.common.data.models.user.Tutor
 import com.nailorsh.repeton.features.tutorsearch.data.TutorSearchRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
